+++ gulp-tasks/workbox/workbox-reference-docs.js
@@ -1,12 +1,12 @@
 const gulp = require('gulp');
 const path = require('path');
 
-const buildReferenceDocs = require('./reference-docs/build-reference-docs');
+const buildReferenceDocs = require('../reference-docs/build-reference-docs');
 
-gulp.task('reference-docs:workbox', () => {
+gulp.task('workbox-reference-docs', () => {
   const GIT_URL = 'https://github.com/GoogleChrome/workbox.git';
   const toolsPath = path.join(
-    __dirname, '..', 'src', 'content', 'en', 'tools'
+    __dirname, '..', '..', 'src', 'content', 'en', 'tools'
   );
   const docPath = path.join(toolsPath, 'workbox', 'reference-docs');
   const jsdocConfPath = path.join(toolsPath, 'workbox', '_jsdoc.conf');
